+++ Source/Extension.Voting/src/adminPage/adminPage.html
@@ -63,6 +63,13 @@
                 </div>
             </div>
 
+            <hr/>
+
+            <div class="row" style="padding-top: 10px;">
+                <h5>Report</h5>
+                <div class="col-sm-12" id="reportContainer"></div>
+            </div>
+
             <div class="row" style="padding-top: 10px;">
                 <div class="col-sm-12" id="errorMessage" v-bind:class="{ hide: showContent }">
                     <p id="votingdisabledmessage">There is no voting available. Please create a new voting.</p>
