+++ Source/Extension.Voting/src/adminPage/adminPage.ts
@@ -1,5 +1,6 @@
 ﻿import { LogExtension } from "../shared/logExtension";
 import { AdminPageController } from "./adminPageController";
+import { ReportPageController } from "../reportPage/reportPageController";
 
 (function () {
     LogExtension.debugEnabled = false;
@@ -8,6 +9,7 @@ import { AdminPageController } from "./adminPageController";
         VSS.ready(() => {
             LogExtension.log("VSS ready");
             new AdminPageController().$mount("#adminPage");
+            new ReportPageController().$mount("#reportContainer");
             $('#appVersion').text(VSS.getExtensionContext().version);
         });
     } catch (ex) {
