+++ Source/Extension.Voting/src/adminPage/adminPageController.ts
@@ -1,10 +1,8 @@
 ﻿import { Voting } from "../entities/voting";
-import { VotingStatus } from "../entities/votingStatus";
 import { AdminPageService } from "./adminPageService";
 import { LogExtension } from "../shared/logExtension";
 import { bsNotify, escapeText } from "../shared/common";
 import * as controls from "VSS/Controls";
-import * as combos from "VSS/Controls/Combos";
 import * as dialogs from "VSS/Controls/Dialogs";
 import * as menus from "VSS/Controls/Menus";
 import * as navigation from "VSS/Controls/Navigation";
@@ -22,19 +20,16 @@ export class AdminPageController extends Vue {
     public levels: string[] = [];
     public userIsAdmin: boolean = true;
     public showContent: boolean = false;
-
-    public created() {
-        document.getElementById("adminPage").classList.remove("hide");
-    }
-
+    
     public mounted() {
         this.adminPageService = new AdminPageService();
-
+        
         this.waitControl = controls.create(statusIndicators.WaitControl, $('#waitContainer'), {
             message: "Loading..."
         });
-
+        
         this.initializeAdminpageAsync();
+        this.$el.classList.remove("hide");
     }
 
     public async addToIncludes(ev) {
